+++ gameio/hoard.cpp
@@ -157,7 +157,7 @@ gameData.hoard.orb.nFrames = cf.ReadShort ();
 gameData.hoard.orb.nWidth = cf.ReadShort ();
 gameData.hoard.orb.nHeight = cf.ReadShort ();
 CalcHoardItemSizes (gameData.hoard.orb);
-fPos = cf.Tell ();
+fPos = (int) cf.Tell ();
 cf.Seek (long (palette.Size () + gameData.hoard.orb.nSize), SEEK_CUR);
 gameData.hoard.goal.nFrames = cf.ReadShort ();
 cf.Seek (fPos, SEEK_SET);
